requirements.txt
altair==4.2.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.0
cycler==0.11.0
entrypoints==0.4
fonttools==4.42.1
gitdb==4.0.10
GitPython==3.1.34
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
networkx==3.1
numpy==1.25.2
packaging==23.1
pandas==2.1.0
Pillow==10.0.0
protobuf==3.20.3
pyarrow==13.0.0
pydeck==0.8.1b0
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.10.2
scikit-learn==1.3.0
scipy==1.11.2
semver==3.0.1
six==1.16.0
sklearn==0.0.post9
smmap==5.0.0
streamlit==1.12.0
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.4
validators==0.22.0
watchdog==3.0.0
zipp==3.16.2
